test: fix existing tests
The run() function in `src/restore.ts` was run twice when a test
is run. Once because `lib/restore` is imported into the test so
it runs automatically, and once because the restore file is
executed from the test.

Removed the wait commands. I think those were there to wait for
the run() function which is imported through `lib/restore` to
complete. Removing the import removes the need to wait.

Merged the two tests into one test. The save action requires
that the restore action was already run once, othewise it
doesn't find the buildcache binary. I think it is better to
execute the restore code in the test, and avoid importing
`lib/restore`. This is also doing the same thing as a real
build would do, instead of running the save code in a "fake"
restored state in the tests.