feat: avoid storing empty/no-change caches + add toggle to skip cache store
Avoid storing the cache if it is empty. This might be because the
job failed to download the cache, or because the job was cancelled
before downloading the cache. In this case, creating a new empty
cache will ruin it for the next job.

Avoid storing the cache if there was no change. This improves the
run time of the job and it also avoids storing multiple identical
caches.

Add option to skip storing of caches. This can be used in jobs that
are only run on PR branches, where the cache will never be used by
anyone.